fix: 删除重复代码；检查文件是否已下载，不重复下载。

+++ lib/transformer.py
@@ -39,9 +39,7 @@ def request(article_id, user_agent):
 
 
 def request_json(article_id, user_agent):
-    article_api_url = f'https://api.zhihu.com/articles/{article_id}'
-    headers = {'user-agent': user_agent}
-    return requests.get(article_api_url, headers=headers).json()
+    return request(article_id, user_agent).json()
 
 
 def preprocess_content(content, download_image, asset_path):
@@ -58,6 +56,8 @@ def preprocess_content(content, download_image, asset_path):
             image_url = matchobj.group(1)
             image_title = image_url.split('/')[-1]
             image_download_path = os.path.join(asset_path, image_title)
+            if os.path.exists(image_download_path):
+                return
             image = requests.get(image_url).content
             with open(image_download_path, 'wb') as image_file:
                 image_file.write(image)
